requirements.txt
fastapi[standard]==0.115.5
python-dotenv